Add integration installers to utils package
